# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Imagine API
  version: 0.1.0
  description: OpenAPI specification for the Imagine API (chi-based)
servers:
  - url: http://localhost:7770
    description: Local API
paths:
...
  /auth/apikey:
    get:
      summary: Generate an API key
      operationId: generateApiKey
      responses:
        "200":
          description: API key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth:
    get:
      summary: Initiate OAuth flow
      operationId: initiateOAuth
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [google, github]
          description: OAuth provider (google or github)
      responses:
        "307":
          description: Redirect to OAuth provider
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth/{provider}:
    post:
      summary: Complete OAuth flow
      operationId: completeOAuth
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
          description: OAuth provider (google or github)
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from OAuth provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State parameter for CSRF protection
      responses:
        "200":
          description: OAuth successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthUserData"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/logout:
    post:
      summary: Logout current session
      operationId: logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    APIKeyResponse:
      type: object
      properties:
        consumer_key:
          type: string
      required: [consumer_key]
    OAuthUserData:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        picture:
          type: string
          format: uri
      required: [email, name, picture]
...